Refactor: Move string formatting to Utils file.
